\name{summary.cglm}

\alias{summary.cglm}

\title{
Summarizes instrumental variable estimation
}

\description{
This is a \code{summary} method for class \code{"cglm"}.
}

\usage{
\method{summary}{cglm}(object, ...)
}

\arguments{
  \item{object}{
an object of class \code{"cglm"}.
} 
 \item{\dots}{
not used.
}
}

\author{
Arvid Sjolander
}

\examples{

##See documentation for cglm.

}
